backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.2
pydantic-settings>=2.0.0
pydantic[email]>=2.5.2
python-dotenv>=1.0.0
user-agents>=2.2.0
python-jose[cryptography]>=3.3.0